Parse UI counters as u64 so wasm32 is not the ceiling
The UI compiles to wasm32, where usize is 32 bits, while the server reports its
counters as 64-bit. Any value above u32::MAX fails deserialization of the whole
response rather than a single field, so a table whose expirationIndex had wrapped
around (see the my-no-sql-sdk expiration index fix) took the entire status page
down with it, and a table over 4GiB of dataSize would have done the same.

API models now hold u64, with their consumers in the components and pages following.
wwwroot rebuilt via ui/build.sh.

+++ ui/src/components/atoms/sparkline.rs
@@ -3,7 +3,7 @@ use dioxus::prelude::*;
 use crate::utils::format_bytes_per_sec;
 
 #[component]
-pub fn Sparkline(values: Vec<usize>, #[props(default)] bytes_label: bool) -> Element {
+pub fn Sparkline(values: Vec<u64>, #[props(default)] bytes_label: bool) -> Element {
     let width: f64 = 200.0;
     let height: f64 = 36.0;
     let count = values.len().max(1) as f64;
