+++ ui/src/models/connections_api_model.rs
@@ -8,11 +8,11 @@ pub struct ConnectionReaderApiModel {
     pub namespace: String,
     pub ip: String,
     #[serde(rename = "incomingPerSecond", default)]
-    pub incoming_per_second: usize,
+    pub incoming_per_second: u64,
     #[serde(rename = "outgoingPerSecond", default)]
-    pub outgoing_per_second: usize,
+    pub outgoing_per_second: u64,
     #[serde(rename = "pendingToSend", default)]
-    pub pending_to_send: usize,
+    pub pending_to_send: u64,
     #[serde(rename = "isNode", default)]
     pub is_node: bool,
 }
@@ -32,21 +32,21 @@ pub struct ConnectionWriterApiModel {
     #[serde(rename = "lastIncomingTime", default)]
     pub last_incoming_time: String,
     #[serde(rename = "reqPerSecond", default)]
-    pub req_per_second: usize,
+    pub req_per_second: u64,
     #[serde(rename = "bytesPerSecond", default)]
-    pub bytes_per_second: usize,
+    pub bytes_per_second: u64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
 pub struct ConnectionsApiModel {
     #[serde(rename = "incomingPerSecond", default)]
-    pub incoming_per_second: usize,
+    pub incoming_per_second: u64,
     #[serde(rename = "outgoingPerSecond", default)]
-    pub outgoing_per_second: usize,
+    pub outgoing_per_second: u64,
     #[serde(rename = "writePayloadsPerSecond", default)]
-    pub write_payloads_per_second: usize,
+    pub write_payloads_per_second: u64,
     #[serde(rename = "writeBytesPerSecond", default)]
-    pub write_bytes_per_second: usize,
+    pub write_bytes_per_second: u64,
     #[serde(default)]
     pub readers: Vec<ConnectionReaderApiModel>,
     #[serde(default)]
